define deploy_current_plan
It runs the deployer on the complete plan. It uses run_planners,
and is therefore able to run method or state machine actions
before it checks for the deployment state.

It returns a mapping from the missions as they currently are to
the deployed missions
