+++ lib/syskit/test/spec.rb
@@ -1,129 +1,14 @@
 # frozen_string_literal: true
 
 require "roby/test/spec"
+require "syskit/test/base"
+require "syskit/test/instance_requirement_planning_handler"
 
 module Syskit
     module Test
-        # Planning handler for #roby_run_planner that handles
-        # InstanceRequirementsTask
-        class InstanceRequirementPlanningHandler
-            # @param [Spec] test
-            def initialize(test)
-                @test = test
-            end
-
-            def start(tasks)
-                @plan = tasks.first.plan
-                @planning_tasks = tasks.map do |t|
-                    unless (planning_task = t.planning_task)
-                        raise ArgumentError, "#{t} does not have a planning task"
-                    end
-
-                    planning_task
-                end
-
-                starting_tasks = @planning_tasks.find_all do |t|
-                    t.start! if t.pending?
-                    t.starting?
-                end
-                return apply_requirements if starting_tasks.empty?
-
-                starting_tasks.each do |t|
-                    t.start_event.on do |_|
-                        apply_requirements if starting_tasks.all?(&:running?)
-                    end
-                end
-            end
-
-            def apply_requirements
-                @plan.syskit_start_async_resolution(
-                    @planning_tasks,
-                    validate_generated_network:
-                        @test.syskit_run_planner_validate_network?,
-                    compute_deployments:
-                        @test.syskit_run_planner_deploy_network?
-                )
-            end
-
-            def finished?
-                if @plan.syskit_has_async_resolution?
-                    return unless @plan.syskit_finished_async_resolution?
-
-                    @plan.syskit_apply_async_resolution_results
-                    return unless @test.syskit_run_planner_stub?
-
-                    root_tasks = @planning_tasks.map(&:planned_task)
-                    stub_network = StubNetwork.new(@test)
-
-                    # NOTE: this is a run-planner equivalent to syskit_stub_network
-                    # we will have to investigate whether we could implement one with
-                    # the other (probably), but in the meantime we must keep both
-                    # in sync
-                    mapped_tasks = @plan.in_transaction do |trsc|
-                        mapped_tasks =
-                            stub_network.apply_in_transaction(trsc, root_tasks)
-                        trsc.commit_transaction
-                        mapped_tasks
-                    end
-
-                    stub_network.remove_obsolete_tasks(mapped_tasks)
-                end
-                @planning_tasks.all?(&:finished?)
-            end
-        end
-        Roby::Test::Spec.roby_plan_with(
-            Component.match.with_child(InstanceRequirementsTask),
-            InstanceRequirementPlanningHandler
-        )
-
-        Roby::Test::ExecutionExpectations.include ExecutionExpectations
-
-        module InstanceRequirementPlanningHandlerOptions
-            def setup
-                @syskit_run_planner_stub = true
-                @syskit_run_planner_validate_network = false
-                @syskit_run_planner_deploy_network = false
-
-                super
-            end
-
-            def syskit_run_planner_stub?
-                @syskit_run_planner_stub
-            end
-
-            attr_writer :syskit_run_planner_stub
-
-            def syskit_run_planner_validate_network?
-                @syskit_run_planner_validate_network
-            end
-
-            attr_writer :syskit_run_planner_validate_network
-
-            def syskit_run_planner_deploy_network?
-                @syskit_run_planner_deploy_network
-            end
-
-            attr_writer :syskit_run_planner_deploy_network
-
-            def syskit_run_planner_with_full_deployment(stub: true)
-                flags = [@syskit_run_planner_deploy_network,
-                         @syskit_run_planner_validate_network,
-                         @syskit_run_planner_stub]
-
-                @syskit_run_planner_deploy_network = true
-                @syskit_run_planner_validate_network = true
-                @syskit_run_planner_stub = stub
-                yield
-            ensure
-                @syskit_run_planner_deploy_network,
-                    @syskit_run_planner_validate_network,
-                    @syskit_run_planner_stub = *flags
-            end
-        end
-
         class Spec < Roby::Test::Spec
             include Test::Base
-            include InstanceRequirementPlanningHandlerOptions
+            include InstanceRequirementPlanningHandler::Options
 
             def setup
                 NetworkGeneration::Engine.on_error = :save
